+++ Examples/Resources/AADEntitlementManagementAccessPackage/2-Update.ps1
@@ -24,16 +24,15 @@ Configuration Example
     {
         AADEntitlementManagementAccessPackage 'myAccessPackage'
         {
-            AccessPackagesIncompatibleWith = @();
-            CatalogId                      = "General";
+            DisplayName           = "Integration Package"
+            Description           = "Integration Tests"
+            CatalogId             = "General"
+            IsHidden              = $True # Updated Property
+            IsRoleScopesVisible   = $True
+            Ensure                = "Present"
             ApplicationId         = $ApplicationId
             TenantId              = $TenantId
             CertificateThumbprint = $CertificateThumbprint
-            Description                    = "Integration Tests";
-            DisplayName                    = "Integration Package";
-            Ensure                         = "Present";
-            IsHidden                       = $True; # Updated Property
-            IsRoleScopesVisible            = $True;
         }
     }
 }